tests/wpt/tests/html/semantics/embedded-content/media-elements/audio_loop_seek_to_eos.html

<!DOCTYPE html>
<html>
<head>
  <title>Seeking to the end of looping audio</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/media.js"></script>
</head>
<body>
<div id="log"></div>
<audio id="a" controls loop></audio>
<script type="text/javascript">

promise_test(async t => {
  const a = document.getElementById("a");
  a.src = getAudioURI("/media/sound_5") + "?" + new Date() + Math.random();
  await a.play();

  // Seek to the end of audio (EOS). However, as audio is looping, it should
  // keep playing after seeking.
  a.currentTime = a.duration;
  await new Promise(r => a.onseeked = r);
  await new Promise(r => a.ontimeupdate = r);
  assert_false(a.paused);
} , "seeking to the end of looping audio");

</script>
</body>
</html>
